requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.6.1
jinja2==3.1.4
python-multipart==0.0.20
python-slugify==8.0.4
psycopg2-binary==2.9.10
markdown==3.7
cryptography==44.0.2
